refactor(backend): consolidate sdk config into libs/sdkconfig
SdkConfig & friends were duplicated across ingest/measure &
libs/measure. The two copies had already drifted: http_sampling_rate
existed only in libs, & since ingest is the only path SDKs fetch
config from, that field never reached a single SDK.

Move the shared code into a new leaf package libs/sdkconfig rather
than into libs/measure. libs/measure spans 30+ files & imports 20
other backend packages, so importing it for SdkConfig pulled go-mail,
posthog-go, semver & libs/posthog into ingest's compile closure.
The new package depends only on uuid, pgx, sqlf & valkey.

Also drops ingest's dead createDefaultConfig & invalidateCache.
Cache mechanics are unchanged here.
